requirements.txt
streamlit
streamlit-webrtc
crewai
langchain
chromadb
beautifulsoup4
openai-whisper
pyttsx3
soundfile
tiktoken
numpy
torch
transformers
huggingface-hub
sentence-transformers
pyaudio
langchain_community
javalang
